refactor: extract deterministic_sync/arrive into deterministic.h
These two functions were duplicated verbatim across mainloop_bwd,
epilogue_bwd, and epilogue_fwd. Move them to a shared header to
eliminate the triple copy. Verified: 5 deterministic cases (plain,
pack_gqa, cat_gqa × aligned/unaligned) all PASS with FWD+BWD
bitwise equality.
